dont download + unpack artifacts if already present

+++ ops/run-cljdoc-api.sh
@@ -5,13 +5,15 @@ set -euo pipefail
 # TODO print instructions if file does not exist
 version=$(cat "$HOME/CLJDOC_VERSION")
 
-echo "Downloading cljdoc v$version"
-# TODO fail if 404
-curl -Ls "https://github.com/martinklepsch/cljdoc/archive/$version.tar.gz" -o "cljdoc-$version.tar.gz"
+if [ ! -d "cljdoc-$version" ]; then
+    echo "Downloading cljdoc v$version"
+    # TODO fail if 404
+    curl -Ls "https://github.com/martinklepsch/cljdoc/archive/$version.tar.gz" -o "cljdoc-$version.tar.gz"
 
-echo "Unpacking archive"
-tar -xvf "cljdoc-$version.tar.gz"
-rm "cljdoc-$version.tar.gz"
+    echo "Unpacking archive"
+    tar -xvf "cljdoc-$version.tar.gz"
+    rm "cljdoc-$version.tar.gz"
+fi
 
 pushd "cljdoc-$version"
